requirements.txt
mcp[cli]>=1.2.0
httpx
python-dateutil